refactor: address PR review feedback on warning cleanup
- table_compute: drop redundant static_cast<uint64_t> around
  INVALID_INDEX_64 (the ull literal is already uint64_t).
- interface_types.h: remove stray comment about where INVALID_INDEX
  is defined.
- timeline_view, track_item: initialize SubscriptionToken members with
  EventManager::InvalidSubscriptionToken instead of the raw
  INVALID_INDEX_64, and drop the now-unused rocprofvis_shared_types.h
  include each file no longer needs.
- utils: move safe_getenv from an anonymous namespace into
  RocProfVis::View (qualifying the one call from the file-local
  display_looks_forwarded helper).
- track: serve kRPVControllerTrackHistogramBucketDensityIndexed from
  GetUInt64 instead of GetDouble (it is a uint64 count, not a double);
  update the two callers (data_provider, controller system tests) to
  rocprofvis_controller_get_uint64.
- table_processor: use uint32_t for the aggregation thread/row locals,
  and declare query offset/limit as uint32_t (atoll -> atol). Removes
  the size_t/int comparison warnings and the redundant row_index casts.

+++ src/controller/src/compute/rocprofvis_controller_table_compute.cpp
@@ -232,7 +232,7 @@ rocprofvis_result_t ComputeTable::Load(const std::string& csv_file)
             {
                 if (csv_row[col].is_null())
                 {
-                    row_data.emplace_back(static_cast<uint64_t>(INVALID_INDEX_64)); // Rows may have empty fields.
+                    row_data.emplace_back(INVALID_INDEX_64); // Rows may have empty fields.
                 }
                 else
                 {
